Create cognitirworkshop
This is some practice that i got to do with Mr. Yu Chen. Super cool guy I got to meet through CFALA. Guess what? He's going to teach one of the courses at USC Analytics this Spring. Guess who's going to jump on the course when it's available :dancing_men: